created scanner overlays with relative and absolute sizes with respect to the camera size
